completed task 6
